requirements.txt
asgiref==3.4.1
dataclasses==0.8
Django==3.1.8
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
typing-extensions==3.10.0.2
